Feat(Business) Add Business Logic calculate distance and needed fuel

+++ FlightPlanning/FlightPlanning.Services.Flights/Dto/AircraftDto.cs
@@ -9,9 +9,9 @@ namespace FlightPlanning.Services.Flights.Dto
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public decimal Speed { get; set; }
-        public decimal FuelCapacity { get; set; }
-        public decimal FuelConsumption { get; set; }
-        public decimal TakeOffEffort { get; set; }
+        public double Speed { get; set; }
+        public double FuelCapacity { get; set; }
+        public double FuelConsumption { get; set; }
+        public double TakeOffEffort { get; set; }
     }
 }
